refactor and update table logic
